WIP: T06 -- Split tests/unit/core/test_command_executor_pipeline.py by theme
