+++ tests/unit/core/CLAUDE.md
@@ -11,6 +11,8 @@
 
 - `set_registry_apps(registry, apps)` — configures a `mock_registry`'s `__contains__`, `app_keys()`, `get_running_apps()`, and `get()` from an `apps`-shaped dict (`dict[str, dict[int, App]]`); use instead of assigning `mock_registry.apps = ...` directly (that attribute no longer exists on the real `AppRegistry`)
 - `make_executor(**kw)` — real `CommandExecutor` with dependencies mocked out
+- `init_executor(queue_max=10)` — real `CommandExecutor` set up for write-pipeline tests (bounded queue, capacity-warning config, real `ready_event`); shared by `test_command_executor_pipeline_queue.py`, `test_command_executor_pipeline_persist.py`, `test_command_executor_pipeline_serve.py`
+- `make_invocation(**kw)` — `ExecutionRecord` for a handler execution, defaults tuned for the write-pipeline tests above
 - `make_mock_cmd_listener(**kw)` — `MagicMock` Listener for `CommandExecutor` tests (side_effect, error_handler)
 - `make_execute_job_cmd(**kw)` — `MagicMock` spec'd to `ExecuteJob` for executor tests
 - `make_bus_service(**kw)`, `make_scheduler_service(**kw)` — service instances bypassing `Resource.__init__`
